refactor(mcp): SqliteStore is now a test-only fixture; drop OnnxEmbedder/download/stats-cli
