--- packages/mcp/src/store/types.ts
@@ -1,12 +1,0 @@
-// Re-export Store interface and related types from @ato-mcp/shared
-// Kept for backwards compatibility; prefer importing directly from @ato-mcp/shared.
-export type {
-  Store,
-  StatsResult,
-  DefinitionRow,
-  ThresholdRow,
-  AnchorRow,
-  CitationRow,
-  DocResult,
-  AnchorGraph,
-} from "@ato-mcp/shared";
