+++ packages/mcp/test/helpers/sqlite-store.ts
